adding db migration

+++ Api/Models/CourseAssignment.cs
@@ -2,8 +2,8 @@ namespace TFL.DevOps.Api.Models;
 
 public class CourseAssignment
 {
-    public long InstructorID {get;set;}
-    public long CourseID {get;set;}
+    public int InstructorID {get;set;}
+    public int CourseID {get;set;}
     
     public Instructor? Instructor {get;set;}
     public Course? Course {get;set;}
