+++ Api/Models/Enrollment.cs
@@ -7,9 +7,9 @@ public enum Grade
 
 public class Enrollment
 {
-    public long EnrollmentId {get;set;}
-    public long CourseId {get;set;}
-    public long StudentID {get;set;}
+    public int EnrollmentId {get;set;}
+    public int CourseId {get;set;}
+    public int StudentID {get;set;}
     public Grade? Grade {get;set;}
 
     public Course? Course {get;set;}
